Changed user profile to swipe between past/future events
