+++ package.json
@@ -25,9 +25,11 @@
     "react-native": "0.79.2",
     "react-native-gesture-handler": "~2.24.0",
     "react-native-get-location": "^5.0.0",
+    "react-native-pager-view": "6.7.1",
     "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "^5.4.1",
     "react-native-screens": "^4.11.1",
+    "react-native-tab-view": "^4.1.2",
     "react-native-web": "^0.20.0",
     "react-native-webview": "^13.13.5"
   },
